requirements.txt
requests==2.26.0
Django==2.2.13
djangorestframework==3.12.4
django-crispy-forms==1.1.0
